Moved `DollarAmount` to common
