+++ kifi-backend/modules/shoebox/app/com/keepit/payments/PaidAccount.scala
@@ -1,41 +1,12 @@
 package com.keepit.payments
 
 import com.keepit.common.db._
-import com.keepit.common.json.KeyFormat
+import com.keepit.common.mail.EmailAddress
 import com.keepit.common.time._
 import com.keepit.model._
-import com.keepit.common.mail.EmailAddress
 import com.keepit.social.BasicUser
-
 import com.kifi.macros.json
 import org.joda.time.DateTime
-import play.api.libs.functional.syntax._
-import play.api.libs.json._
-
-@json
-case class DollarAmount(cents: Int) extends Ordered[DollarAmount] {
-  def compare(that: DollarAmount) = cents compare that.cents
-  def +(other: DollarAmount): DollarAmount = DollarAmount(cents + other.cents)
-  def -(other: DollarAmount): DollarAmount = DollarAmount(cents - other.cents)
-  def *(x: Int) = DollarAmount(cents * x)
-  def max(other: DollarAmount): DollarAmount = DollarAmount(cents max other.cents)
-  def min(other: DollarAmount): DollarAmount = DollarAmount(cents min other.cents)
-  override def toString = toDollarString
-  def toDollarString: String = if (cents < 0) "-" + (-this).toDollarString else "$%d.%02d".format(cents / 100, cents % 100)
-
-  def unary_- = DollarAmount(-1 * cents)
-
-  def toCents: Int = cents
-}
-
-object DollarAmount {
-  def cents(cents: Int): DollarAmount = DollarAmount(cents)
-  def dollars(dollars: Int): DollarAmount = DollarAmount(dollars * 100)
-
-  val ZERO = DollarAmount(0)
-
-  val formatAsCents: Format[DollarAmount] = (__ \ 'cents).format[Int].inmap(DollarAmount.cents, _.toCents)
-}
 
 @json
 case class SimpleAccountContactInfo(who: BasicUser, enabled: Boolean)
